rough-cut/python/requirements.txt
httpx>=0.27.0
budoux>=0.6.0
pyyaml>=6.0
kanjize>=1.0.0
